+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to the "everforest" extension will be documented in this fil
 
 - Remove `everforest.darkSelection` and `everforest.lightSelection` customization options.
 - Always use the default selection colors in all workbench styles.
+- Remove `everforest.darkCursor` and `everforest.lightCursor` customization options.
+- Always use the default cursor colors in all workbench styles.
 
 ## 0.3.0